access control and permission blocking

+++ prf-todos-api/api/resources.py
@@ -1,22 +1,40 @@
 from api.models import Todo, TodoItem, User
 from api.schemas import TodoListItemSchema, TodoListSchema, UserSchema
 from flask import Blueprint, jsonify, request
+from flask_login import current_user, login_required, login_user, logout_user
 from flask_restful import Api, Resource
 from marshmallow import ValidationError
 from prftodosapi.extensions import db
+from werkzeug.security import check_password_hash
 
 blueprint = Blueprint("api", __name__, url_prefix="/api")
 api = Api(blueprint)
 
 
 class TodoItemResource(Resource):
-    # @checkPermission()
+    @login_required
     def get(self, todo_id):
+        curr_user_id = int(current_user.get_id())
+        curr_user_todos = db.session.scalars(
+            db.select(Todo.id).filter_by(user_id=curr_user_id)).all()
+
+        if not (todo_id in curr_user_todos):
+            return "Todo not found", 404
+
         todo_items = db.session.scalars(
             db.select(TodoItem).filter_by(todo_id=todo_id)).all()
+
         return TodoListItemSchema().dump(todo_items, many=True)
 
+    @login_required
     def post(self, todo_id):
+        curr_user_id = int(current_user.get_id())
+        curr_user_todos = db.session.scalars(
+            db.select(Todo.id).filter_by(user_id=curr_user_id)).all()
+
+        if not (todo_id in curr_user_todos):
+            return "Todo not found", 404
+
         try:
             data = TodoListItemSchema().load(request.json)
             todo_item = TodoItem(**data)
@@ -26,9 +44,21 @@ class TodoItemResource(Resource):
             return err.messages, 400
         return None
 
+    @login_required
     def patch(self, todo_id, item_id):
+        curr_user_id = int(current_user.get_id())
+        curr_user_todos = db.session.scalars(
+            db.select(Todo.id).filter_by(user_id=curr_user_id)).all()
+
+        if not (todo_id in curr_user_todos):
+            return "Todo not found", 404
+
         schema = TodoListItemSchema(partial=True, exclude=['todo_id'])
-        todo_item = TodoItem.query.get_or_404(item_id)
+        todo_item: TodoItem = TodoItem.query.get_or_404(item_id)
+
+        if not (todo_item.todo_id in curr_user_todos):
+            return "Item not found", 404
+
         data = schema.load(request.json)
         for key, value in data.items():
             if hasattr(todo_item, key):
@@ -36,20 +66,33 @@ class TodoItemResource(Resource):
         db.session.commit()
         return None
 
+    @login_required
     def delete(self, todo_id, item_id):
+        curr_user_id = int(current_user.get_id())
+        curr_user_todos = db.session.scalars(
+            db.select(Todo.id).filter_by(user_id=curr_user_id)).all()
+
+        if not (todo_id in curr_user_todos):
+            return "Todo not found", 404
+
         item = TodoItem.query.get_or_404(item_id)
+
+        if not (item.todo_id in curr_user_todos):
+            return "Item not found", 404
+
         db.session.delete(item)
         db.session.commit()
         return None
 
 
 class TodoResource(Resource):
-
+    @login_required
     def get(self, user_id):
         todos = db.session.scalars(
             db.select(Todo).filter_by(user_id=user_id)).unique().all()
         return TodoListSchema(many=True).dump(todos)
 
+    @login_required
     def post(self):
         try:
             data = TodoListSchema().load(data=request.json)
@@ -60,7 +103,15 @@ class TodoResource(Resource):
             return err.messages, 400
         return None
 
+    @login_required
     def patch(self, todo_id):
+        curr_user_id = int(current_user.get_id())
+        curr_user_todos = db.session.scalars(
+            db.select(Todo.id).filter_by(user_id=curr_user_id)).all()
+
+        if not (todo_id in curr_user_todos):
+            return "Todo not found", 404
+
         schema = TodoListSchema(partial=True)
         todo = Todo.query.get_or_404(todo_id)
         data = schema.load(data=request.json)
@@ -70,7 +121,15 @@ class TodoResource(Resource):
         db.session.commit()
         return None
 
+    @login_required
     def delete(self, todo_id):
+        curr_user_id = int(current_user.get_id())
+        curr_user_todos = db.session.scalars(
+            db.select(Todo.id).filter_by(user_id=curr_user_id)).all()
+
+        if not (todo_id in curr_user_todos):
+            return "Todo not found", 404
+
         todo = Todo.query.get_or_404(todo_id)
         db.session.delete(todo)
         db.session.commit()
@@ -78,7 +137,7 @@ class TodoResource(Resource):
 
 
 class UserResource(Resource):
-
+    @login_required
     def get(self, user_id):
         user = User.query.get_or_404(user_id)
         return jsonify(user)
@@ -94,6 +153,7 @@ class UserResource(Resource):
             return err.messages
         return None
 
+    @login_required
     def patch(self, user_id):
         schema = UserSchema(partial=True)
         user = User.query.get_or_404(user_id)
@@ -104,10 +164,28 @@ class UserResource(Resource):
         db.session.commit()
         return None
 
+    @login_required
     def delete(self, user_id):
         user = User.query.get_or_404(user_id)
         db.session.delete(user)
         db.session.commit()
+
+
+class AuthLoginResource(Resource):
+    def post(self):
+        data = UserSchema().load(request.json)
+        user = db.session.execute(
+            db.select(User).filter_by(username=data["username"])).scalar_one()
+        if not user or not check_password_hash(user.password_hash, data["password"]):
+            return 'Login fail', 404
+        login_user(user)
+        return "Login success", 200
+
+
+class AuthLogoutResource(Resource):
+    def post(self):
+        logout_user()
+        return "logout success", 200
 
 
 api.add_resource(TodoItemResource, "/todos/<int:todo_id>/items",
@@ -115,3 +193,6 @@ api.add_resource(TodoItemResource, "/todos/<int:todo_id>/items",
 api.add_resource(TodoResource, "/users/<int:user_id>/todos",
                  "/todos", "/todos/<int:todo_id>")
 api.add_resource(UserResource, "/users", "/users/<int:user_id>")
+
+api.add_resource(AuthLoginResource, "/login")
+api.add_resource(AuthLogoutResource, "/logout")
